refactor(adapt): split anchor_adapt.rs into per-concern dir module
Move-only carve of the 1,751-line anchor_adapt.rs into
adapt/anchor_adapt/{mod,adapter,attributes,extract,render,handler_override,tests}.rs.

The facade mod.rs preserves every crate::anchor_adapt::* / self::anchor_adapt::*
path at its current visibility; cross-bucket private fns bumped to pub(super)
only. Zero behavior change: all four snapshot suites byte-identical, 1007 unit
+ 24 module tests green, clippy --bin qedgen --tests -D warnings clean, fmt clean.

Part of the v3.0 cleanup (monolith-split thread).
