+++ crates/qedgen/src/adapt/anchor_adapt/render.rs
@@ -1,0 +1,162 @@
+use super::*;
+
+pub(super) fn render_spec(model: &ProgramModel) -> String {
+    let mut s = String::new();
+    s.push_str("// Generated by `qedgen adapt`. Fill in the TODOs to make this verifiable.\n");
+    if let (Some(primary_source), Some(entry_module)) = (&model.primary_source, &model.entry_module)
+    {
+        s.push_str(&format!(
+            "// Source: {} (program mod: `{}`)\n\n",
+            primary_source.display(),
+            entry_module,
+        ));
+    }
+    s.push_str(&format!("spec {}\n\n", to_pascal_case(&model.name)));
+
+    s.push_str("// TODO: replace with the actual lifecycle of your program.\n");
+    s.push_str("type State\n");
+    s.push_str("  | Init\n");
+    s.push_str("  | Active\n\n");
+
+    match model.errors.as_ref() {
+        Some(info) if !info.variants.is_empty() => {
+            let source_path = info
+                .source_path
+                .as_ref()
+                .map(|p| p.display().to_string())
+                .unwrap_or_else(|| "<unknown>".to_string());
+            s.push_str(&format!(
+                "// Error variants discovered in {} (`#[error_code] pub enum {}`).\n",
+                source_path, info.enum_name,
+            ));
+            s.push_str("type Error\n");
+            for variant in &info.variants {
+                s.push_str(&format!("  | {}\n", variant));
+            }
+            s.push('\n');
+        }
+        Some(info) => {
+            let source_path = info
+                .source_path
+                .as_ref()
+                .map(|p| p.display().to_string())
+                .unwrap_or_else(|| "<unknown>".to_string());
+            s.push_str(&format!(
+                "// Found `#[error_code] pub enum {}` in {} but it has no variants.\n",
+                info.enum_name, source_path,
+            ));
+            s.push_str("// TODO: list domain errors raised by the handlers below.\n");
+            s.push_str("type Error\n");
+            s.push_str("  | InvalidArgument\n\n");
+        }
+        None => {
+            s.push_str("// TODO: list domain errors raised by the handlers below.\n");
+            s.push_str("// (No `#[error_code]` enum found in the program's source.)\n");
+            s.push_str("type Error\n");
+            s.push_str("  | InvalidArgument\n\n");
+        }
+    }
+
+    for handler in &model.handlers {
+        render_handler(&mut s, handler);
+        s.push('\n');
+    }
+
+    s
+}
+
+fn render_handler(s: &mut String, entry: &HandlerModel) {
+    match &entry.shape {
+        HandlerShape::Inline => {
+            s.push_str(&format!(
+                "/// `{}` — inline body in the `#[program]` mod\n",
+                entry.name
+            ));
+        }
+        HandlerShape::FreeFn => {
+            s.push_str(&format!("/// `{}` — free-fn forwarder\n", entry.name));
+        }
+        HandlerShape::Method { impl_type } => {
+            s.push_str(&format!(
+                "/// `{}` — method on `{}`\n",
+                entry.name, impl_type
+            ));
+        }
+        HandlerShape::Unrecognized { reason } => {
+            s.push_str(&format!(
+                "/// `{}` — UNRECOGNIZED forwarder ({})\n",
+                entry.name, reason
+            ));
+            s.push_str(
+                "/// TODO: classify this handler manually. The body may use a\n\
+                 ///       custom dispatcher or a shape the adapter doesn't\n\
+                 ///       cover yet.\n",
+            );
+        }
+        HandlerShape::SourceWalk => {
+            s.push_str(&format!(
+                "/// `{}` — discovered via source-walk\n",
+                entry.name
+            ));
+        }
+    }
+    if let Some(path) = &entry.source_path {
+        s.push_str(&format!("/// discovered at: {}\n", path.display()));
+    }
+    if let Some(accounts) = &entry.accounts_type {
+        s.push_str(&format!(
+            "/// accounts struct: `{}` (see `#[derive(Accounts)]`)\n",
+            accounts
+        ));
+    }
+
+    // qedspec only accepts `//` line comments (no `/* */`), so arg-type
+    // fallback notes go inside the body, not the signature.
+    s.push_str(&format!("handler {}", entry.name));
+    let mut unknown_args: Vec<&str> = Vec::new();
+    for arg in &entry.args {
+        match &arg.qedspec_type {
+            Some(ty) => s.push_str(&format!(" ({} : {})", arg.name, ty)),
+            None => {
+                // Unknown type → U64 placeholder so the spec parses;
+                // surfaced in a body comment.
+                s.push_str(&format!(" ({} : U64)", arg.name));
+                unknown_args.push(arg.name.as_str());
+            }
+        }
+    }
+    s.push_str(" : State.Init -> State.Init {\n");
+    if !unknown_args.is_empty() {
+        s.push_str(&format!(
+            "  // TODO: refine arg types — could not map {} from Rust source (likely generic / Vec / Option).\n",
+            unknown_args
+                .iter()
+                .map(|a| format!("`{}`", a))
+                .collect::<Vec<_>>()
+                .join(", "),
+        ));
+    }
+    s.push_str("  // TODO: auth <signer>\n");
+    s.push_str("  // TODO: accounts { ... }\n");
+    s.push_str("  // TODO: requires\n");
+    s.push_str("  // TODO: effect { ... }\n");
+    s.push_str("}\n");
+}
+
+/// snake_case → PascalCase (program mod name `my_escrow` → spec name
+/// `MyEscrow`).
+pub(super) fn to_pascal_case(s: &str) -> String {
+    let mut out = String::with_capacity(s.len());
+    let mut upper_next = true;
+    for ch in s.chars() {
+        if ch == '_' {
+            upper_next = true;
+        } else if upper_next {
+            out.push(ch.to_ascii_uppercase());
+            upper_next = false;
+        } else {
+            out.push(ch);
+        }
+    }
+    out
+}
